+++ public/main/context/TimerResource.ts
@@ -1,5 +1,10 @@
 import {IResource} from "./SparkmaxContext";
 
+/**
+ * Resource used by {@link SparkmaxContext} based on interval.
+ *
+ * {@link TimerResource} periodically (`ms`) starts action (`processor`).
+ */
 export class TimerResource implements IResource {
   private intervalId: any;
   private currentProcessing?: Promise<void>;
@@ -11,29 +16,45 @@ export class TimerResource implements IResource {
     this.start();
   }
 
+  /**
+   * Waits until current processing is completed and stops timer
+   */
   public destroy(): Promise<void> {
     this.destroyed = true;
     return this.stop();
   }
 
+  /**
+   * Waits until current processing is completed and stops timer.
+   * Processing can be resumed later.
+   */
   public pause(): Promise<void>|undefined {
     this.checkAlive();
 
     return this.stop();
   }
 
+  /**
+   * Resumes processing.
+   */
   public resume(): void {
     this.checkAlive();
 
     this.start();
   }
 
+  /**
+   * Changes target device
+   */
   public setOwner(device: string): void {
     this.checkAlive();
 
     this.device = device;
   }
 
+  /**
+   * Starts timer
+   */
   private start(): void {
     this.checkAlive();
 
@@ -45,6 +66,9 @@ export class TimerResource implements IResource {
     }, this.ms);
   }
 
+  /**
+   * Stops timer
+   */
   private stop(): Promise<void> {
     clearInterval(this.intervalId);
     return this.currentProcessing || Promise.resolve();
