+++ src/components/fields/SwitchParamField.tsx
@@ -2,9 +2,9 @@ import {Switch} from "@blueprintjs/core";
 import {isFunction} from "lodash";
 import * as React from "react";
 import {useCallback} from "react";
-import {IParamSourceProps} from "../param-source";
+import {IConfigParamProps} from "../config-param-props";
 
-interface IProps extends IParamSourceProps {
+interface IProps extends IConfigParamProps {
   className?: string;
   label?: string | ((checked: boolean) => string);
   inverted?: boolean;
