+++ src/store/ram-config-param-rules.ts
@@ -24,6 +24,7 @@ export const createRamConfigParamContext = (state: IApplicationState): IRamConfi
  * Returns registry for {@link IConfigParamRule}s suitable to work with stored in RAM values.
  */
 export const getRamConfigParamRule = mapRuleRegistry(getConfigParamRule, [
+  // Implement validations for in-RAM parameters
   mapRule((rule) => ({
     ...rule,
     validate: (ctx: IRamConfigParamContext) => {
@@ -37,7 +38,7 @@ export const getRamConfigParamRule = mapRuleRegistry(getConfigParamRule, [
         return message;
       }
 
-      // Validate the latest value in RAM
+      // Analyze the latest response that came from the server
       const param = getDeviceParamOrDefault(querySelectedDeviceCurrentConfig(state), rule.id);
       const {lastResponse} = param;
       if (lastResponse && lastResponse.status === ParamStatus.Invalid) {
@@ -46,11 +47,16 @@ export const getRamConfigParamRule = mapRuleRegistry(getConfigParamRule, [
 
       return warning;
     },
-    getMessage(ctx: IRamConfigParamContext): Message | undefined {
+    /**
+     * Return the latest validation message
+     * @param ctx
+     */
+    getMessage: (ctx: IRamConfigParamContext): Message | undefined => {
       const param = getDeviceParamOrDefault(querySelectedDeviceCurrentConfig(ctx.getState()), rule.id);
       return param.message;
-    }
+    },
   })),
+  // Implement additional validations for CAN ID field
   mapRule(ConfigParam.kCanID, (rule) => ({
     ...rule,
     validate: (ctx: IRamConfigParamContext) => {
@@ -59,13 +65,17 @@ export const getRamConfigParamRule = mapRuleRegistry(getConfigParamRule, [
 
       const device = querySelectedDevice(state)!;
       if (value === 0 && device.uniqueId !== 0) {
+        // Generate error if device is not configured
         return Message.error(MESSAGE_NOT_CONFIGURED_DEVICE);
       } else if (without(queryConnectedDevicesByCanId(state, value), device).length > 0) {
+        // Generate error if CAN ID is not unique on the bus
         return Message.error(substitute(MESSAGE_NOT_UNIQUE_CAN_ID, {id: String(value)}));
       } else if (value === 0) {
+        // Generate warning if device is a single device having CAN ID = 0
         return Message.warning(MESSAGE_CAN_0);
       }
 
+      // Run other validations
       return rule.validate(ctx);
     },
   })),
